requirements.txt
# Core packages directly imported in aqua-app
streamlit==1.50.0
streamlit-authenticator==0.1.5
pandas==2.2.3
numpy==1.26.4
matplotlib==3.10.7
seaborn==0.13.2
folium==0.20.0
geopandas==1.1.1
pillow==11.3.0
plotly==6.3.1
PyYAML==6.0.3
pyprojroot==0.3.0
psycopg2-binary==2.9.11
sshtunnel==0.3.2
pydeck==0.9.1
SQLAlchemy==1.4.35
# time==0.1.1
# streamlit-spinner==0.1.2

# # Essential dependencies for the above packages
# # geopandas dependencies
# shapely==2.1.2
# pyproj==3.7.2
# pyogrio==0.11.1
# # folium dependencies
# branca==0.8.2
# Jinja2==3.1.6
# # pandas dependencies
# python-dateutil==2.9.0.post0
# pytz==2025.2
# # matplotlib dependencies
# kiwisolver==1.4.9
# cycler==0.12.1
# fonttools==4.60.1
# contourpy==1.3.3
# pyparsing==3.2.5
# # plotly dependencies
# tenacity==9.1.2
# # streamlit dependencies
# click==8.3.0
# pyarrow==16.1.0
# tornado==6.5.2
# tzdata==2025.2
# # General utilities
# certifi==2025.10.5
# urllib3==2.5.0
# requests==2.32.5
# packaging==25.0
# MarkupSafe==3.0.3
# six==1.17.0

# # host on google cloud
# google-cloud-aiplatform
# google-cloud-logging
